docker/fastapi/requirements.txt
fastapi
sse-starlette
uvicorn[standard]
asyncio
python-dotenv
redis